# optimize permute! and invpermute!