--- a/src/PooledArrays.jl
+++ b/src/PooledArrays.jl
@@ -297,6 +297,16 @@ function Base.reverse(x::PooledArray)
     PooledArray(RefArray(reverse(x.refs)), x.invpool, x.pool, x.refcount)
 end
 
+function Base.permute!(x::PooledArray, p::AbstractVector{T}) where T<:Integer
+    permute!(x.refs, p)
+    return x
+end
+
+function Base.invpermute!(x::PooledArray, p::AbstractVector{T}) where T<:Integer
+    invpermute!(x.refs, p)
+    return x
+end
+
 function Base.permute!!(x::PooledArray, p::AbstractVector{T}) where T<:Integer
     Base.permute!!(x.refs, p)
     return x