After https://github.com/JuliaLang/julia/pull/44941 `permute!` and `invpermute!` no longer call the `Base.permute!!` and `Base.invpermute!!` methods. This ensures that `permute!` and `invpermute!` stay optimized for `PooledArray`s, otherwise they get a suboptimal fallback (see benchmark below).

```julia
julia> using BenchmarkTools, PooledArrays, Random

julia> perm = randperm(10_000);

julia> @btime permute!($v, $perm) setup=(v=PooledArray(rand(1:1000, 10_000)));
  9.800 μs (2 allocations: 39.11 KiB)

julia> @btime ($v .= $v[$perm]) setup=(v=PooledArray(rand(1:1000, 10_000))); # current permute! fallback
  10.800 μs (4 allocations: 39.17 KiB)

julia> @btime invpermute!($v, $perm) setup=(v=PooledArray(rand(1:1000, 10_000)));
  11.900 μs (2 allocations: 39.11 KiB)

julia> @btime ($v[$perm] = $v) setup=(v=PooledArray(rand(1:1000, 10_000))); # current invpermute! fallback
  103.400 μs (4 allocations: 39.17 KiB)
```

The improvement is much more substantial for `invpermute!`, but I imagine it makes sense to have both methods fully optimized. I was actually surprised that the `permute!` fallback was already almost optimal. EDIT: that was just due to it calling the optimized `getindex(v::PooledArray, i::AbstractVector)` and `copyto!(dst::PooledArray, src::PooledArray)` methods, whereas `invpermute!` has to call `setindex!`, which is more costly.